refactor(model)!: Flatten module
After having some time using it, the nested nature just bothered me.
