+++ crates/core/src/runtime/variable.rs
@@ -1,7 +1,7 @@
 use std::fmt;
 
 use crate::error::{Error, Result};
-use crate::model::find::Path;
+use crate::model::Path;
 use crate::model::Scalar;
 use crate::model::{ValueCow, ValueView};
 
@@ -95,8 +95,8 @@ impl fmt::Display for Variable {
 mod test {
     use super::*;
 
-    use crate::model::value::ValueViewCmp;
     use crate::model::Object;
+    use crate::model::ValueViewCmp;
 
     use super::super::RuntimeBuilder;
     use super::super::StackFrame;
