+++ crates/lib/src/stdlib/blocks/if_block.rs
@@ -2,7 +2,7 @@ use std::fmt;
 use std::io::Write;
 
 use liquid_core::error::ResultLiquidExt;
-use liquid_core::model::{value::ValueViewCmp, ValueView};
+use liquid_core::model::{ValueView, ValueViewCmp};
 use liquid_core::parser::BlockElement;
 use liquid_core::parser::TagToken;
 use liquid_core::Expression;
